gecko/dom/tv/test/mochitest/test_tv_scan_channels_completed.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test channel scanning complete for TV API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript" src="./head.js"></script>
<script type="application/javascript">

function runTest() {
  ok('tv' in navigator, "navigator.tv should exist.");

  var isScannedEventFired = false;
  var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL('tv_chrome_script.js'));

  navigator.tv.getTuners().then(
    function(aTuners) {
      ok(aTuners.length > 0, "Got at least 1 tuner.");

      aTuners[0].getSources().then(
        function(aSources) {
          ok(aSources.length > 0, "Got at least 1 source.");
          var source = aSources[0];

          source.onscanningstatechanged = function(aEvent) {
            if (aEvent.state === 'scanned') {
              if (!isScannedEventFired) {
                isScannedEventFired = true;
                gScript.sendAsyncMessage('trigger-channel-scan-complete',
                                         { tunerId: aTuners[0].id,
                                           sourceType: source.type });
              }
              info("Received channel scanned event.");
              ok(aEvent.channel, "Scanned channel should be set.");
            } else if (aEvent.state === 'completed') {
              ok(isScannedEventFired, "Received channel scanning completed event after channel scanned event.");
              gScript.destroy();
              SimpleTest.finish();
            }
          };

          // TODO Bug 1088818 - Modify the behavior of channel scanning.
          source.startScanning({}).then(
            function() {
              gScript.sendAsyncMessage('trigger-channel-scanned',
                                       { tunerId: aTuners[0].id,
                                         sourceType: source.type });
            },
            function(aError) {
              ok(false, "Error occurred when starting scanning: " + aError);
              gScript.destroy();
              SimpleTest.finish();
            }
          );
        },
        function(aError) {
          ok(false, "Error occurred when getting sources: " + aError);
          gScript.destroy();
          SimpleTest.finish();
        }
      );
    },
    function(aError) {
      ok(false, "Error occurred when getting tuners: " + aError);
      gScript.destroy();
      SimpleTest.finish();
    }
  );
}

SimpleTest.expectAssertions(0, 2);
SimpleTest.waitForExplicitFinish();
prepareTest(runTest);

</script>
</pre>
</body>
</html>
